+++ src/components/OverviewItem/OverviewItem.tsx
@@ -2,12 +2,20 @@ import React from "react";
 import { CLASSPREFIX as eccgui } from "../../configuration/constants";
 
 export interface OverviewItemProps extends React.HTMLAttributes<HTMLDivElement> {
-    children?: React.ReactNode
-    className?: string;
+    /**
+     * Displays the element using reduced height and less white space inside.
+     */
     densityHigh?: boolean;
+    /**
+     * Add a bit white space around the element.
+     */
     hasSpacing?: boolean;
 }
 
+/**
+ * This element can include all basic information and actions to give an overview about the item.
+ * Mainly used in items list or to create basic widgets.
+ */
 function OverviewItem({
     children,
     className = '',