+++ src/components/Tag/Tag.tsx
@@ -1,22 +1,48 @@
 import React from 'react';
 import {
     Tag as BlueprintTag,
-    ITagProps as IBlueprintTagPRops
+    TagProps as BlueprintTagProps
 } from "@blueprintjs/core";
+import Color from "color";
 import { CLASSPREFIX as eccgui } from "../../configuration/constants";
 
-interface TagProps extends IBlueprintTagPRops {
-    emphasis?: "stronger" | "strong" | "normal" | "weak" | "weaker";
-    small?: boolean;
+interface TagProps extends Omit<
+    BlueprintTagProps,
     // currently not supported
-    active?: never;
-    fill?: never;
-    large?: never;
-    multiline?: never;
-    intent?: never;
-    rightIcon?: never;
+    "active" |
+    "fill" |
+    "large" |
+    "multiline" |
+    "intent" |
+    "rightIcon" |
+    "round"
+> {
+    // own properties
+
+    /**
+     * Sets the background color of a tag, depends on the `Color` object provided by the
+     * [npm color module](https://www.npmjs.com/package/color) v3. You can use it with
+     * all allowed [CSS color values](https://developer.mozilla.org/de/docs/Web/CSS/color_value).
+     *
+     * The front color is set automatically, so the tag label is always readable.
+     */
+    backgroundColor?: Color | string;
+
+    /**
+     * visual appearance and "thickness" of the tag
+     */
+    emphasis?: "stronger" | "strong" | "normal" | "weak" | "weaker";
+    /**
+     * display tag in a small version
+     */
+    small?: boolean;
+
     // deprecated
-    emphasized?: never; // use minimal=false plus emphasis="stronger"
+
+    /**
+     * **deprecated**, use `minimal=false` plus `emphasis="stronger"`
+     */
+    emphasized?: never;
 }
 
 function Tag({
@@ -25,8 +51,21 @@ function Tag({
     emphasis = "normal",
     minimal = true,
     small = false,
+    backgroundColor,
     ...otherProps
 }: TagProps) {
+    otherProps['interactive'] = otherProps.interactive ?? !!otherProps.onClick ? true : false;
+    if (!!backgroundColor) {
+        const additionalStyles = otherProps.style ?? {};
+        const color = Color(backgroundColor);
+        otherProps['style'] = {
+            ...additionalStyles,
+            ...{
+                backgroundColor: color.rgb().toString(),
+                color: color.isLight() ? "#000" : "#fff",
+            }
+        }
+    }
     return (
         <BlueprintTag
             {...otherProps}