+++ CHANGELOG.md
@@ -3,7 +3,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [unreleased]
+## [Unreleased]
+
+### Added
+
+* `Tag` element got new property for `backgroundColor`
+* Styles for footnotes and task lists, rendered by Markdown GFM parser.
+
+### Fixed
+* Add missing import to `components/Spinner/Spinner.tsx`. Make gui-elements compatible with typescript `^4.5.2`.
+* Add bottom white space in tables in content block elements.
+* `fullWidth` on/off display of `TextField` is now working like expected
+
+### Changed
+* `TextField` elements are using `fullWidth=true` by default
+* `SearchField` uses now by default `"operation-search"` as `leftIcon`
 
 ### Deprecated